Add Java sample code to README (#27)
Hello, crux is such a lovely library! However, I'm using it from Java instead of Kotlin, which adds a little complexity. I figured out some code that works, and I thought it would be useful to everyone else if that example code was included in the README. Hopefully the sample makes sense!